Updated user classes to user package and added audit fields to user models.

+++ amcart-facades/src/com/nagarro/amcart/facade/user/data/UserData.java
@@ -1,27 +1,27 @@
 package com.nagarro.amcart.facade.user.data;
 
 public class UserData {
-	private Long id;
-	private String name;
+    private Integer id;
+    private String name;
 
-	public Long getId() {
-		return id;
-	}
+    public Integer getId() {
+        return id;
+    }
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	@Override
-	public String toString() {
-		return "UserData [id=" + id + ", name=" + name + "]";
-	}
+    @Override
+    public String toString() {
+        return "UserData [id=" + id + ", name=" + name + "]";
+    }
 }
